convert Badge to typescript
